+++ src/index.ts
@@ -6,5 +6,6 @@
 export * from "./crypto.js";
 export * from "./encoding.js";
 export * from "./model.js";
+export * from "./physical-binding.js";
 export * from "./replay.js";
 export * from "./session.js";